refactor Log messages and include GL

+++ Sparky/src/maths/maths.h
@@ -11,10 +11,10 @@ namespace sparky
 {
 	namespace maths
 	{
-		inline float toRadians(float degrees) { return degrees * (M_PI / 180.0f); };
+		inline double toRadians(double degrees) { return degrees * (M_PI / 180.0f); };
 
-		inline float toDegrees(float radians) { return radians * (180.0f / M_PI); }
+		inline double toDegrees(double radians) { return radians * (180.0f / M_PI); }
 
-		inline int sign(float value) { return (value > 0) - (value < 0); }
+		inline int sign(double value) { return (value > 0) - (value < 0); }
 	}
 }
